refactor(tests): restructure test suite to mirror source code layout
Split monolithic test_commands.py and test_workspace.py into focused
files under tests/commands/ and tests/utils/ to match the src/ow/
module structure established in the previous commit.

New test files:
- commands/test_status.py, test_rebase.py, test_create.py,
  test_update.py, test_prune.py
- utils/test_templates.py, test_drift.py, test_display.py
- utils/test_git.py, test_config.py, test_resolver.py (moved)

220 tests pass. Zero tests lost or duplicated.
